express-openapi-validator => 3.16.4
There was a problem following references where the filename of a request and
response ref were different. This turned out to be a bug in
express-openapi-validator. The maintainer pushed out a fix.

This commit moves this library to the version containing the fix, it
also switches us to using components for the query request ref.