QueryLabels refactoring
In the response to the request to add or remove a label to/from a query,
we need to return the updated query object.

In order to achieve this in a way that didn't lead to a tonne of
duplicate code, I have refactored the 'addEmbeds' function into a new
helper.

Also, the controllers for adding a label to a query, and removing a
label from a query were, unsurprisingly, almost identical. I have
therefore replaced them with a single function.

Includes:

- addLabelToQuery & removeLabelFromQuery replaced with addRemove
- The relevant router now passes an argument to addRemove to indicate
what operation it should carry out
- New helpers/queries.js helper
- New units tests where appropriate
- Modified unit tests where appropriate