Query bulk update & bulk query/label
This commit adds the ability for the API to accept requests to bulk
update queries and perform bulk attachment/detachment of a label
from multiple queries.

Includes:
- New PUT /queries route which allows an array of query objects to be
passed, which will be treated as query updates. The response is an array
of updated query objects
- Modification to POST|DELETE /query/{query_id}/label/{label_id}. query_id can
now be a comma delimited list of query IDs. This allows us to assign or
remove a label from multiple queries in one call. The response is an
array of the affected queries
- New queryrequests & queryresponses schemas to define the payloads
mentioned above
- A new queries.updateBulk controller which handles updates to multiple queries
- Modifications to querylabel.addRemove controller to switch it to
always treat query_id as a potential array. The handling is identical
either way and it will always return an array of updated queries
- New and modified unit tests to cover the above