Add send of websocket messages upon message events
This commit adds the sending of websocket messages to clients when
certain message related events happen.

Sends updated state when:

- Message or upload is created
- Message is updated
- Message or upload is deleted

It includes a new collection of middleware that are called after the
route controllers that carries out "side-effects" that are required by
the endpoint. For example, if a message is sent, an updated copy of the
query metadata is sent to all clients via websocket.