TokenCache => AuthCache plus storing role
The TokenCache was just storing refresh tokens in-memory. This also
seems to be a logical place to store a user's role code, which is
obtained as part of the refresh token generation and return flow.

Consequently, TokenCache has been renamed to AuthCache and we're now
storing the refresh token and the user's role. This will enable the API
role lookup to not hit the DB.

Includes:

- Rename TokenCache to AuthCache
- Renamed AuthCache methods appropriately
- As part of refresh token => JWT flow, we now look up the user object,
get their role code and store it along with the refresh token.
- Modified unit tests with new names