Add user role validation on API routes
This commits adds the validation of the requesting user's role when a
request to an API route is made. Each route has a list of roles required
to access it, this is checked prior to hitting the route controller.

Consequently, when the JWT validation passes, we look up the user
described in the JWT and add them to the request object, this makes them
available to be checked against in the subsequent validation middleware.

Includes:
- Fetch and store a user with a valid JWT
- Add role requirements to each route definition, a request that fails a
validation returns 401