Fix response definition and fix WS upload support
The update of express-openapi-validator appears to have added extra
strictness which highlighted an error in the upload response definition.
This commit fixes this bug.

It also fixes a bug in the preparation of the wsData object that is
passed to the middleware, it should be passed the query object, this was
not the case.