Fix minor bugs relating to unseen counts
A couple of bugs existing when a query is created. Since a query is
created then immediately after the first message / upload is sent, we
need the client's mostRecentSeen value to be available. This commit
sends this value to all connected clients upon creation of a new query.

Includes:
- Created new side-effect middleware, mostRecentSeenToAll, which sends
the most recent seen values to all connected clients
- Added mostRecentSeenToAll to queries POST chain
- Added queryUnseenCountsToClients & folderCountsToClients to uploads
POST chain. This was to fix a bug whereby unseen counts were not being
updated upon upload
- Added creation of queryuser entry upon query creation, so it can be
returned to the user ASAP