Per-instance upload directories
Something that wasn't apparent until we had multiple prod deployments
for testing was that all instances were using the same directory for
uploads. This is not a great idea... So we now have per-instance
subdirectories. They're based on instance DB schema name, so
predictable, we maybe should add some randomness to them, but for now
they're sufficient I think.