Add start script
This commit moves the encryption key under which user emails are stored
into something that is pasted in when the app starts.

- Remove KEY property from app environment in
ecosystem.config.js.template
- Create start.sh, a script that accepts arguments containing the
instance name and environment, then prompts the user to paste the
encryption key, then starts the app, passing the key into the app's
environment.